+++ lindenmayer/src/svg.rs
@@ -3,7 +3,7 @@ use std::fs::File;
 use crate::draw::*;
 
 pub fn draw_svg_utf8() -> Vec<u8> {
-    let (after, _) = develop_system();
+    let (after, _) = crate::develop_system();
 
     let mut v = vec![];
 
@@ -14,7 +14,7 @@ pub fn draw_svg_utf8() -> Vec<u8> {
 
 #[allow(unused)]
 pub fn draw_to_file() {
-    let (after, iters) = develop_system();
+    let (after, iters) = crate::develop_system();
 
     let mut file = File::create(&format!("plant_{:02}.svg", iters).to_string()).unwrap();
     draw(&after, 0.0, 20.0, 10.0, &mut file);
